Refactor interfaces and base classes
