+++ src/api.js
@@ -10,18 +10,18 @@ const { Address, Transaction } = BitcoinSource
  * Rest API interface that all supported chains must implement
  * @interface
  */
-export interface Api {
+export interface IApiMinimal {
   /**
    * Retrieves a given address' history.
    * @throws {ApiError} if the request cannot be completed.
-   * @memberof Api
+   * @memberof IApiMinimal
    */
   getAddress(address: Address): Promise<Object>;
 
   /**
    * Retrieves a given address' balance in satoshis.
    * @throws {ApiError} if the request cannot be completed.
-   * @memberof Api
+   * @memberof IApiMinimal
    */
   getBalance(address: Address): Promise<number>;
 
@@ -29,46 +29,30 @@ export interface Api {
    * Sends a transaction for broadcasting.
    * @returns the resulting transaction id
    * @throws {ApiError} if the request cannot be completed.
-   * @memberof Api
+   * @memberof IApiMinimal
    */
   sendTransaction(transaction: Transaction): Promise<TransactionId>;
-
-  /**
-   * Retrieves a block from its hash or height.
-   * @param {string | number} hashOrHeight Hash or height of the block
-   * @throws {ApiError} if the request cannot be completed.
-   * @memberof Api
-   */
-  getBlock(hashOrHeight: string | number): Promise<Object>;
 
   /**
    * Retrieves the hash of a block from its height.
    * @param {*} height Block height
    * @throws {ApiError} if the request cannot be completed.
-   * @memberof Api
+   * @memberof IApiMinimal
    */
   getBlockHash(height: number): Promise<string>;
 
   /**
    * Retrives the hash of the latest block.
    * @throws {ApiError} if the request cannot be completed.
-   * @memberof Api
+   * @memberof IApiMinimal
    */
   getLastBlockHash(): Promise<string>;
-
-  /**
-   * Retrives a hex-formatted block given its hash or height.
-   * @param {string | number} hashOrHeight Hash or height of the block
-   * @throws {ApiError} if the request cannot be completed.
-   * @memberof Api
-   */
-  getRawBlock(hashOrHeight: string | number): Promise<string>;
 
   /**
    * Retrieves a JSON-formatted transaction from its hash
    * @param {string} txId Transaction hash
    * @throws {ApiError} if the request cannot be completed.
-   * @memberof Api
+   * @memberof IApiMinimal
    */
   getTransaction(txId: string): Promise<Object>;
 
@@ -76,7 +60,7 @@ export interface Api {
    * Retrieves a hex-formatted transaction given its hash
    * @param {*} txId Transaction hash
    * @throws {ApiError} if the request cannot be completed.
-   * @memberof Api
+   * @memberof IApiMinimal
    */
   getRawTransaction(txId: string): Promise<Object>;
 
@@ -84,7 +68,7 @@ export interface Api {
    * Retrieves a given address' unspent outputs (UTXO set).
    * @param {Address} address Address whose UTXOs to retrieve
    * @throws {ApiError} if the request cannot be completed.
-   * @memberof Api
+   * @memberof IApiMinimal
    */
   getUtxos(address: Address): Promise<Array<Txo>>;
 
@@ -92,7 +76,31 @@ export interface Api {
    * Gets a transaction output given an output id.
    * @param {OutputId} outputId Transaction id and output index
    * @throws {ApiError} if the request cannot be completed.
-   * @memberof Api
+   * @memberof IApiMinimal
    */
   getTxo(outputId: OutputId): Promise<Txo>;
+}
+
+/**
+ * Rest API interface that all defines a complete implementation
+ * @interface
+ */
+export interface IApi extends IApiMinimal {
+
+  /**
+   * Retrieves a json-formatted block from its hash or height.
+   * @param {string | number} hashOrHeight Hash or height of the block
+   * @throws {ApiError} if the request cannot be completed.
+   * @memberof IApi
+   */
+  getBlock(hashOrHeight: string | number): Promise<Object>;
+
+  /**
+   * Retrives a hex-formatted block given its hash or height.
+   * @param {string | number} hashOrHeight Hash or height of the block
+   * @throws {ApiError} if the request cannot be completed.
+   * @memberof IApi
+   */
+  getRawBlock(hashOrHeight: string | number): Promise<string>;
+
 }
